Fix the path to STL argument
The python script did not work because the first argument of the stl_to_gltf fucntion on the stl2gltf.py was not used on the script. Instead the variable `path_to_stl` is used. The argument name was changed to `path_to_stl`. Now the script works as expected.